pyproject.toml
[project]
name = "gdocs-chat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bs4>=0.0.2",
    "chromadb>=1.0.20",
    "ipython>=9.5.0",
    "langchain>=0.3.27",
    "langchain-chroma>=0.2.5",
    "langchain-community>=0.3.29",
    "langchain-core>=0.3.75",
    "langchain-openai>=0.3.32",
    "langgraph>=0.6.6",
    "langsmith>=0.4.21",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "pypdf>=6.0.0",
    "python-dotenv>=1.1.1",
    "matplotlib>=3.8.0",
    "pillow>=10.0.0",
]
